emissions use double instead of float, better naming

+++ src/ext/java/org/opentripplanner/ext/digitransitemissions/DefaultEmissionsService.java
@@ -15,7 +15,7 @@ public class DefaultEmissionsService implements Serializable, EmissionsService {
   }
 
   @Override
-  public Float getEmissionsForItinerary(Itinerary itinerary) {
+  public Double getEmissionsForItinerary(Itinerary itinerary) {
     Optional<DigitransitEmissionsService> service = repository.retrieveEmissionsService();
     if (service.isPresent()) {
       return service.get().getEmissionsForItinerary(itinerary);
